services/delivery/requirements.txt
azure-cosmos>=4.7.0
azure-storage-blob>=12.21.0
azure-identity>=1.17.0
jinja2>=3.1.4
pytest>=8.3.0
